requirements.txt
flask
tensorflow
numpy
pillow
matplotlib
seaborn
pandas
scikit-learn
